Separate Notify request helpers from our own APIs
The new helper is a less surprising one to include in the spec for
the new "bulk" rake task, since it's no longer coupled to the type
of spec being run.
