+++ spec/spec_helper.rb
@@ -19,6 +19,7 @@ RSpec.configure do |config|
   config.use_transactional_fixtures = true
   config.include AuthenticationHelpers, type: :request
   config.include RequestHelpers, type: :request
+  config.include NotifyRequestHelpers, type: :request
   config.include FactoryBot::Syntax::Methods
   config.include ActiveSupport::Testing::TimeHelpers
