+++ spec/support/request_helpers.rb
@@ -1,12 +1,4 @@
 module RequestHelpers
-  def stub_notify
-    allow_any_instance_of(DeliveryRequestService)
-      .to receive(:provider_name).and_return("notify")
-
-    body = {}.to_json
-    stub_request(:post, /fake-notify/).to_return(body: body)
-  end
-
   def create_subscriber_list(overrides = {})
     params = { title: "Example", tags: {}, links: {} }.merge(overrides)
     post "/subscriber-lists", params: params.to_json, headers: json_headers
@@ -68,23 +60,8 @@ module RequestHelpers
     expect(response.status).to eq(expected_status)
   end
 
-  def expect_an_email_was_sent
-    request_data = nil
-    expectation = ->(request) { request_data = data(request.body) }
-    expect(a_request(:post, /fake-notify/).with(&expectation)).to have_been_made.at_least_once
-    request_data
-  end
-
-  def expect_an_email_was_not_sent
-    expect(a_request(:post, /fake-notify/)).not_to have_been_made
-  end
-
   def data(body = response.body)
     JSON.parse(body).deep_symbolize_keys
-  end
-
-  def clear_any_requests_that_have_been_recorded!
-    WebMock::RequestRegistry.instance.reset!
   end
 
   def json_headers
